+++ common/src/NodePool/TraceNode.cpp
@@ -21,121 +21,100 @@
  */
 
 #include "TraceNode.h"
+#include "common.h"
 #include "header.h"
+#include "json/value.h"
+#include <cstdint>
+#include <memory>
+
 namespace PP {
 namespace NodePool {
-TraceNode::~TraceNode() {}
 
-void TraceNode::clearAttach() {
-  // empty the json value
-  if (!this->_value.empty())
-    this->_value.clear(); // Json::Value();
+TraceNode& TraceNode::Reset(NodeID id) {
+  id_ = id;
 
-  if (!this->context_.empty())
-    this->context_.clear();
+  root_id_ = parent_id_ = next_ = E_INVALID_NODE;
 
-  if (!this->_endTraceCallback.empty())
-    this->_endTraceCallback.clear();
-}
+  parent_start_time_ = 0;
+  trace_start_time_ = 0;
 
-void TraceNode::initId(const NodeID& id) { this->id_ = id; }
+  reference_count_ = 0;
+  root_node_extra_ptr_ = nullptr;
 
-void TraceNode::AddChildTraceNode(WrapperTraceNodePtr& child) {
-  std::lock_guard<std::mutex> _safe(this->mlock);
-  assert(id_ != child->id_);
-
-  if (this->last_child_id_ != E_INVALID_NODE) {
-    child->sibling_id_ = this->last_child_id_;
+  if (!this->user_optional_setting_func_.empty()) {
+    this->user_optional_setting_func_.clear();
   }
 
-  this->last_child_id_ = child->id_;
+  value_ptr_ = std::unique_ptr<Json::Value>(new Json::Value());
 
-  child->parent_id_ = this->id_;
-  child->root_id_ = this->root_id_;
-  child->root_start_time = this->root_start_time;
-  child->parent_start_time = this->start_time;
+  return *this;
 }
 
-void TraceNode::AddChildTraceNode(TraceNode& child) {
-  std::lock_guard<std::mutex> _safe(this->mlock);
-  assert(id_ != child.id_);
+void TraceNode::StartTrace() { trace_start_time_ = get_unix_time_ms(); }
 
-  if (this->last_child_id_ != E_INVALID_NODE) {
-    child.sibling_id_ = this->last_child_id_;
-  }
-  this->last_child_id_ = child.id_;
-
-  child.parent_id_ = this->id_;
-  child.root_id_ = this->root_id_;
-  child.root_start_time = this->root_start_time;
-  child.parent_start_time = this->start_time;
-}
-
-void TraceNode::EndTimer() {
+void TraceNode::EndTrace() {
   uint64_t end_time = get_unix_time_ms();
-  this->cumulative_time += (end_time - this->start_time);
-}
+  this->expired_time += (end_time - this->trace_start_time_);
 
-void TraceNode::WakeUpTimer() { this->start_time = get_unix_time_ms(); }
+  this->AddAnnotation(":E", this->expired_time);
+  this->AddAnnotation(":S", this->trace_start_time_ - this->parent_start_time_);
 
-void TraceNode::EndSpan() {
-  this->AddTraceDetail(":E", this->cumulative_time);
-  this->AddTraceDetail(":S", this->start_time);
   if (this->set_exp_) {
-    this->AddTraceDetail("EA", 1);
+    this->AddAnnotation("EA", 1);
   }
 }
 
-void TraceNode::EndSpanEvent() {
-  this->AddTraceDetail(":E", this->cumulative_time);
-  this->AddTraceDetail(":S", this->start_time - this->parent_start_time);
+void TraceNode::BindParentTrace(WrapperTraceNodePtr& parent) {
+  parent_start_time_ = parent->trace_start_time_;
+  parent_id_ = parent->id_;
+  depth_ = parent->depth_ + 1;
+  root_id_ = parent->root_id_;
 }
 
-void TraceNode::StartTimer() {
-  uint64_t time_in_ms = get_unix_time_ms();
-  this->start_time = time_in_ms;
-  this->root_start_time = time_in_ms;
-}
-
-void TraceNode::parseOpt(std::string key, std::string value) {
+void TraceNode::parseUserOption(std::string key, std::string value) {
   pp_trace(" [%d] add opt: key:%s value:%s", id_, key.c_str(), value.c_str());
   if (key == "TraceMinTimeMs") {
     int64_t min = std::stoll(value);
     auto cb = [=]() -> bool {
       pp_trace("checkOpt:  [%d] TraceMinTimeMs:%" PRIu64 " cumulative_time:%" PRIu64 "", this->id_,
-               min, this->cumulative_time);
-      if ((int64_t)this->cumulative_time >= min)
+               min, this->expired_time);
+      if ((int64_t)this->expired_time >= min) {
         return true;
-      return false;
+      } else {
+        return false;
+      }
     };
-    this->_endTraceCallback.push_back(cb);
+
+    this->user_optional_setting_func_.push_back(cb);
   } else if (key == "TraceOnlyException") {
     auto cb = [=]() -> bool { return this->set_exp_; };
 
-    this->_endTraceCallback.push_back(cb);
+    this->user_optional_setting_func_.push_back(cb);
   }
 }
 
-bool TraceNode::checkOpt() {
+bool TraceNode::runUserOptionFunc() {
   bool ret = true;
-  for (auto& cb : this->_endTraceCallback) {
+  for (auto& cb : this->user_optional_setting_func_) {
     if ((ret = cb()) == true)
       return ret;
   }
   return ret;
 }
 
-void TraceNode::setOpt(const char* opt, va_list* args) {
+void TraceNode::setNodeUserOption(const char* opt, va_list* args) {
   const char* var = opt;
 
   while (var != nullptr) {
     const char* delimit = strchr(var, ':');
     if (delimit == nullptr) {
-      this->parseOpt(std::string(var), "");
+      // key:value format
+      this->parseUserOption(std::string(var), "");
     } else {
+      // command format
       std::string key(var, delimit - var);
       std::string value(delimit + 1);
-      this->parseOpt(key, value);
+      this->parseUserOption(key, value);
     }
     var = va_arg(*args, const char*);
   }