+++ common/src/NodePool/test_node.cc
@@ -2,6 +2,7 @@
 #include "json/value.h"
 #include <gtest/gtest.h>
 #include "header.h"
+#define COMMON_DEBUG
 #include "NodePool/PoolManager.h"
 #include <condition_variable>
 #include <thread>
@@ -21,7 +22,7 @@ std::mutex cv_m;
 std::condition_variable cv;
 NodeID rootId = E_ROOT_NODE;
 
-// note: as it known, there may leak some node
+// note: as it known, there may leak some nodes
 void func() {
   std::unique_lock<std::mutex> lk(cv_m);
   cv.wait(lk);
@@ -70,90 +71,29 @@ TEST(node, multipleThread) {
 void test_opt(TraceNode& node, const char* opt, ...) {
   va_list args;
   va_start(args, opt);
-  node.setOpt(opt, &args);
+  node.setNodeUserOption(opt, &args);
   va_end(args);
 }
 
 TEST(node, opt) {
 
-  TraceNode& node = nodePool_.Take();
+  TraceNode& node = nodePool_.GetNode();
 
   test_opt(node, "TraceMinTimeMs:23", "TraceOnlyException", nullptr);
 
-  node.cumulative_time = 22;
-  EXPECT_FALSE(node.checkOpt());
-  node.cumulative_time = 23;
-  EXPECT_TRUE(node.checkOpt());
+  node.expired_time = 22;
+  EXPECT_FALSE(node.runUserOptionFunc());
+  node.expired_time = 23;
+  EXPECT_TRUE(node.runUserOptionFunc());
 
-  node.cumulative_time = 0;
+  node.expired_time = 0;
   node.set_exp_ = true;
-  EXPECT_TRUE(node.checkOpt());
+  EXPECT_TRUE(node.runUserOptionFunc());
   node.set_exp_ = false;
-  EXPECT_FALSE(node.checkOpt());
+  EXPECT_FALSE(node.runUserOptionFunc());
 
-  nodePool_.ReturnNode(node);
-  // EXPECT_TRUE(PoolManager::getInstance().NoNodeLeak());
-}
-
-static std::string span;
-void capture(const char* msg) {
-  pp_trace("capture:%s", msg);
-  span = std::string(msg);
-}
-//./bin/TestCommon --gtest_filter=node.pinpoint_start_traceV1
-TEST(node, pinpoint_start_traceV1) {
-  pinpoint_set_agent("tcp:127.0.0.1:9999", 0, -1, 7000);
-  register_span_handler(capture);
-  NodeID root, child1;
-  root = pinpoint_start_trace(E_ROOT_NODE);
-  child1 = pinpoint_start_traceV1(root, "TraceMinTimeMs:23", nullptr);
-  pinpoint_add_clue(child1, "name", "Take1sec", E_LOC_CURRENT);
-  sleep(1);
-  pinpoint_end_trace(child1);
-  check_tracelimit(-1);
-  check_tracelimit(0);
-  check_tracelimit(time(nullptr));
-  child1 = pinpoint_start_traceV1(root, "TraceOnlyException", nullptr);
-  pinpoint_add_clue(child1, "name", "Exception", E_LOC_CURRENT);
-  pinpoint_add_exception(child1, "xxxxxxxxx");
-  pinpoint_end_trace(child1);
-
-  child1 = pinpoint_start_traceV1(root, "TraceMinTimeMs:2000", nullptr);
-  pinpoint_add_clue(child1, "name", "TraceMinTimeMs:2000", E_LOC_CURRENT);
-  sleep(1);
-
-  {
-    NodeID child = pinpoint_start_traceV1(child1, "TraceMinTimeMs:23", nullptr);
-    pinpoint_add_clue(child1, "name", "childFromTraceMinTimeMs:23-1", E_LOC_CURRENT);
-    pinpoint_end_trace(child);
-    child = pinpoint_start_traceV1(child1, "TraceMinTimeMs:23", nullptr);
-    pinpoint_add_clue(child1, "name", "childFromTraceMinTimeMs:23-2", E_LOC_CURRENT);
-    pinpoint_end_trace(child);
-    child = pinpoint_start_traceV1(child1, "TraceMinTimeMs:23", nullptr);
-    pinpoint_add_clue(child1, "name", "childFromTraceMinTimeMs:23-3", E_LOC_CURRENT);
-    pinpoint_end_trace(child);
-  }
-
-  pinpoint_end_trace(child1);
-
-  child1 = pinpoint_start_traceV1(root, "TraceOnlyException", nullptr);
-  pinpoint_add_clue(child1, "name", "NoException", E_LOC_CURRENT);
-  pinpoint_end_trace(child1);
-
-  child1 = pinpoint_start_traceV1(root, "TraceMinTimeMs:-23", nullptr);
-  pinpoint_add_clue(child1, "name", "TraceMinTimeMs:-23", E_LOC_CURRENT);
-  pinpoint_end_trace(child1);
-
-  pinpoint_end_trace(root);
-  pp_trace("span:%s", span.c_str());
-  EXPECT_TRUE(span.length() > 0);
-  EXPECT_TRUE(span.find("Take1sec") != span.npos);
-  EXPECT_TRUE(span.find("Exception") != span.npos);
-  EXPECT_TRUE(span.find("TraceMinTimeMs:2000") == span.npos);
-  EXPECT_TRUE(span.find("NoException") == span.npos);
-
-  EXPECT_TRUE(span.find("childFromTraceMinTimeMs:23-3") == span.npos);
-  EXPECT_TRUE(span.find("childFromTraceMinTimeMs:23-2") == span.npos);
+  nodePool_.ReturnNode(node.id_);
+  EXPECT_TRUE(nodePool_.NoNodeLeak());
 }
 
 TEST(node, leak_node) {
@@ -170,61 +110,6 @@ TEST(node, leak_node) {
 
   EXPECT_EQ(count, nodePool_.freeNodesCount());
   show_status();
-}
-std::set<std::string> removed_keys = {":E", ":S"};
-void Remove(Json::Value& v) {
-  Json::Value::Members mem = v.getMemberNames();
-  for (auto iter = mem.begin(); iter != mem.end(); iter++) {
-    if (removed_keys.find(*iter) != removed_keys.end()) {
-      v.removeMember(*iter);
-    }
-
-    if (*iter == "calls") {
-      for (long i = 0; i < v[*iter].size(); i++) {
-        Remove(v["calls"][(int)i]);
-      }
-    }
-  }
-
-  return;
-}
-
-bool check_span_order(std::string& i1, std::string& i2) {
-  Json::Value v_i1, v_i2;
-  Json::Reader reader;
-  reader.parse(i1, v_i1, false);
-  reader.parse(i2, v_i2, false);
-  Remove(v_i1);
-  Remove(v_i2);
-  pp_trace("%s", v_i1.toStyledString().c_str());
-  pp_trace("%s", v_i2.toStyledString().c_str());
-  return v_i1.toStyledString() == v_i2.toStyledString();
-}
-
-TEST(node, call_order) {
-  register_span_handler(capture);
-  NodeID root, child1, child2;
-  root = pinpoint_start_trace(E_ROOT_NODE);
-  pinpoint_add_clue(root, "name:", "root", E_LOC_CURRENT);
-  child1 = pinpoint_start_trace(root);
-  pinpoint_add_clue(child1, "name:", "child1", E_LOC_CURRENT);
-  child2 = pinpoint_start_trace(child1);
-  pinpoint_add_clue(child2, "name:", "child2", E_LOC_CURRENT);
-  child2 = pinpoint_start_trace(child1);
-  pinpoint_add_clue(child2, "name:", "child3", E_LOC_CURRENT);
-  pinpoint_end_trace(child2);
-  child2 = pinpoint_start_trace(child1);
-  pinpoint_add_clue(child2, "name:", "child4", E_LOC_CURRENT);
-  pinpoint_end_trace(child2);
-  child2 = pinpoint_start_trace(child1);
-  pinpoint_add_clue(child2, "name:", "child5", E_LOC_CURRENT);
-  pinpoint_end_trace(child2);
-  pinpoint_end_trace(child1);
-  pinpoint_end_trace(root);
-
-  std::string exp =
-      R"({":E":0,":FT":7000,":S":1710408777521,"calls":[{":E":0,":S":0,"calls":[{"name:":"child2"},{":E":0,":S":0,"name:":"child3"},{":E":0,":S":0,"name:":"child4"},{":E":0,":S":0,"name:":"child5"}],"name:":"child1"}],"name:":"root"})";
-  EXPECT_TRUE(check_span_order(span, exp));
 }
 
 TEST(node, tons_of_nodes_01) {
@@ -243,11 +128,11 @@ TEST(node, tons_of_nodes_01) {
   NodeID child1 = root;
   for (int i = 0; i < 1000; i++) {
     NodeID child = pinpoint_start_trace(child1);
-    mark_current_trace_status(root, E_OFFLINE);
+    change_trace_status(root, E_OFFLINE);
     pinpoint_end_trace(child);
     child1 = child;
   }
-  mark_current_trace_status(root, E_TRACE_BLOCK);
+  change_trace_status(root, E_TRACE_BLOCK);
   pinpoint_end_trace(root);
 }
 // HACK here reproduce a bug in win11, not found any solution
@@ -332,11 +217,14 @@ TEST(node, free_when_add) {
   auto usedNode = getUsedNode();
   NodeID root;
   auto make_it_busy = [&]() {
-    auto w_root = nodePool_.GetWrapperNode();
-    root = w_root->id_;
-    auto w_child = nodePool_.GetWrapperNode();
-    w_root->AddChildTraceNode(w_child);
-    w_child->AddTraceDetail("E_ROOT_NODE", 234);
+    auto& w_root = nodePool_.GetNode();
+    root = w_root.id_;
+    auto& w_child = nodePool_.GetNode();
+
+    WrapperTraceNodePtr parent_ptr = w_root;
+    w_child.BindParentTrace(parent_ptr);
+
+    w_child.AddAnnotation("E_ROOT_NODE", 234);
     std::this_thread::sleep_for(std::chrono::seconds(2));
   };
   std::thread t(make_it_busy);