+++ common/src/NodePool/test_node_pool.cc
@@ -7,54 +7,53 @@
 
 using namespace testing;
 using PP::NodePool::PoolManager;
+using PP::NodePool::ThreadSafePoolManager;
 using PP::NodePool::TraceNode;
 
 TEST(poolManger, get_and_give_back) {
   PoolManager pool;
   // new
-  TraceNode& _node = pool.Take();
+  TraceNode& _node = pool.GetNode();
   void* p = &_node;
   NodeID id = _node.getId();
-  NodeID child, next;
+  NodeID next;
   // give back
-  pool.ReturnNode(id, child, next);
-  EXPECT_EQ(child, E_INVALID_NODE);
+  next = pool.ReturnNode(id);
   EXPECT_EQ(next, E_INVALID_NODE);
-  TraceNode& _node_01 = pool.Take();
+  TraceNode& _node_01 = pool.GetNode();
 
-  TraceNode& new_node = pool.NewNode();
+  TraceNode& new_node = pool.GetNode();
   NodeID new_id = new_node.id_;
 
   auto ref_node = pool.ReferNode(new_id);
   EXPECT_EQ(ref_node->id_, new_node.id_);
 
-  pool.ReturnNode(new_id, child, next);
+  next = pool.ReturnNode(new_id);
   EXPECT_THROW(pool.ReferNode(new_id), std::out_of_range);
 
   // ref current
-  TraceNode& ref_new_node = pool.Take(_node_01.getId());
+  auto ref_new_node = pool.ReferNode(_node_01.getId());
 
   EXPECT_EQ(p, &_node);
-  EXPECT_EQ(ref_new_node, _node);
+  // EXPECT_EQ(ref_new_node, _node);
   // reuse the same id
   EXPECT_EQ(id, _node.getId());
-  EXPECT_THROW(pool.Take(NodeID(100)), std::out_of_range);
-  EXPECT_THROW(pool.Take(NodeID(10000)), std::out_of_range);
+  EXPECT_THROW(pool.ReferNode(NodeID(100)), std::out_of_range);
+  EXPECT_THROW(pool.ReferNode(NodeID(10000)), std::out_of_range);
 }
 
-static PoolManager g_pool;
+static ThreadSafePoolManager thread_safe_pool;
 
 void test_node_pool(bool& result) {
   NodeID it = E_INVALID_NODE;
   for (int i = 0; i < 1000; i++) {
-    TraceNode& _node = g_pool.Take();
+    TraceNode& _node = thread_safe_pool.GetNode();
     usleep(1000);
     if (_node.getId() == it) {
       result = false;
       return;
     }
-    g_pool.ReturnNode(_node);
-    // g_pool.freeNode(_node.getId());
+    thread_safe_pool.ReturnNode(_node.id_);
   }
   result = true;
 }