+++ common/CHANGES.md
@@ -1,3 +1,8 @@
+## v0.6.0 
+- refactor span json protocol
+- add async api
+- data structure change: multiple map to list
+
 ## v0.5.3 2024-07-30
 - support unittest in macos